// src/api/types/git_ref.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GitRefKind } from "./git_ref_kind";

export type GitRef = { name: string, full_ref: string, kind: GitRefKind, target_hash: string, upstream: string | null, ahead: number, behind: number, is_head: boolean, date: bigint, };